move workflow subscription data classes to common lib
