+++ pstatus-graphql-ktor/src/main/kotlin/gov/cdc/ocio/processingstatusapi/mutations/DataStreamTopErrorsNotificationSubscriptionMutationService.kt
@@ -4,34 +4,13 @@ import com.expediagroup.graphql.generator.annotations.GraphQLDescription
 import com.expediagroup.graphql.server.operations.Mutation
 import gov.cdc.ocio.processingstatusapi.ServiceConnection
 import gov.cdc.ocio.processingstatusapi.exceptions.ResponseException
-import gov.cdc.ocio.processingstatusapi.mutations.models.NotificationSubscriptionResult
 import gov.cdc.ocio.processingstatusapi.mutations.response.SubscriptionResponse
-import gov.cdc.ocio.types.model.NotificationType
+import gov.cdc.ocio.types.model.WorkflowSubscription
+import gov.cdc.ocio.types.model.WorkflowSubscriptionResult
 import io.ktor.client.request.*
 import io.ktor.http.*
 import kotlinx.coroutines.runBlocking
 import kotlinx.serialization.Serializable
-
-
-/**
- * DataStream Subscription for digest counts and top errors.
- *
- * @param dataStreamId String
- * @param dataStreamRoute String
- * @param jurisdiction String
- * @param cronSchedule String
- * @param emailAddresses List<String>
- */
-@Serializable
-data class DataStreamTopErrorsNotificationSubscription(
-    val dataStreamId: String,
-    val dataStreamRoute: String,
-    val jurisdiction: String,
-    val cronSchedule: String,
-    val notificationType: NotificationType,
-    val emailAddresses: List<String>?,
-    val webhookUrl: String?
-)
 
 /**
  * DataStream UnSubscription data class used for unsubscribing from the db for digest counts and the top errors and
@@ -64,31 +43,15 @@ class DataStreamTopErrorsNotificationSubscriptionMutationService(
     @GraphQLDescription("Subscribe data stream top errors lets you subscribe to get notifications for top data stream errors and its frequency during an upload")
     @Suppress("unused")
     fun subscribeDataStreamTopErrorsNotification(
-        dataStreamId: String,
-        dataStreamRoute: String,
-        jurisdiction: String,
-        cronSchedule: String,
-        notificationType: NotificationType,
-        emailAddresses: List<String>?,
-        webhookUrl: String?
-    ): NotificationSubscriptionResult {
+        subscription: WorkflowSubscription
+    ): WorkflowSubscriptionResult {
         val url = workflowServiceConnection.buildUrl("subscribe/dataStreamTopErrorsNotification")
 
         return runBlocking {
             val result = runCatching {
                 val response = workflowServiceConnection.client.post(url) {
                     contentType(ContentType.Application.Json)
-                    setBody(
-                        DataStreamTopErrorsNotificationSubscription(
-                            dataStreamId,
-                            dataStreamRoute,
-                            jurisdiction,
-                            cronSchedule,
-                            notificationType,
-                            emailAddresses,
-                            webhookUrl
-                        )
-                    )
+                    setBody(subscription)
                 }
                 return@runCatching SubscriptionResponse.ProcessNotificationResponse(response)
             }
@@ -112,7 +75,7 @@ class DataStreamTopErrorsNotificationSubscriptionMutationService(
     @Suppress("unused")
     fun unsubscribesDataStreamTopErrorsNotification(
         subscriptionId: String
-    ): NotificationSubscriptionResult {
+    ): WorkflowSubscriptionResult {
         val url = workflowServiceConnection.buildUrl("/unsubscribe/dataStreamTopErrorsNotification")
 
         return runBlocking {
