+++ pstatus-graphql-ktor/src/main/kotlin/gov/cdc/ocio/processingstatusapi/mutations/UploadDigestCountsSubscriptionMutationService.kt
@@ -1,37 +1,16 @@
 package gov.cdc.ocio.processingstatusapi.mutations
 
-import gov.cdc.ocio.processingstatusapi.mutations.models.NotificationSubscriptionResult
 import com.expediagroup.graphql.generator.annotations.GraphQLDescription
 import com.expediagroup.graphql.server.operations.Mutation
 import gov.cdc.ocio.processingstatusapi.ServiceConnection
 import gov.cdc.ocio.processingstatusapi.exceptions.ResponseException
 import gov.cdc.ocio.processingstatusapi.mutations.response.SubscriptionResponse
+import gov.cdc.ocio.types.model.WorkflowSubscription
+import gov.cdc.ocio.types.model.WorkflowSubscriptionResult
 import io.ktor.client.request.*
 import io.ktor.http.*
 import kotlinx.coroutines.runBlocking
 import kotlinx.serialization.Serializable
-
-
-/**
- * Daily Upload Digest Counts Subscription data class.
- *
- * @property numDaysAgoToRun Long
- * @property dataStreamIds List<String>
- * @property dataStreamRoutes List<String>
- * @property jurisdictions List<String>
- * @property cronSchedule String
- * @property emailAddresses List<String>
- * @constructor
- */
-@Serializable
-data class UploadDigestCountsSubscription(
-    val numDaysAgoToRun: Long,
-    val dataStreamIds: List<String>,
-    val dataStreamRoutes: List<String>,
-    val jurisdictions: List<String>,
-    val cronSchedule: String,
-    val emailAddresses: List<String>
-)
 
 /**
  * Daily Upload Digest Counts UnSubscription data class which is serialized back and forth which is in turn used for
@@ -65,29 +44,15 @@ class UploadDigestCountsSubscriptionMutationService(
     @GraphQLDescription("Subscribe daily digest counts lets you get notifications with the counts of all jurisdictions for a given set of data streams after the prescribed time to run is past")
     @Suppress("unused")
     fun subscribeUploadDigestCounts(
-        numDaysAgoToRun: Long,
-        dataStreamIds: List<String>? = emptyList(),
-        dataStreamRoutes: List<String>? = emptyList(),
-        jurisdictions: List<String>? = emptyList(),
-        cronSchedule: String,
-        emailAddresses: List<String>
-    ): NotificationSubscriptionResult {
+        subscription: WorkflowSubscription
+    ): WorkflowSubscriptionResult {
         val url = workflowServiceConnection.buildUrl("/subscribe/uploadDigestCounts")
 
         return runBlocking {
             val result = runCatching {
                 val response = workflowServiceConnection.client.post(url) {
                     contentType(ContentType.Application.Json)
-                    setBody(
-                        UploadDigestCountsSubscription(
-                            numDaysAgoToRun,
-                            dataStreamIds ?: emptyList(),
-                            dataStreamRoutes ?: emptyList(),
-                            jurisdictions ?: emptyList(),
-                            cronSchedule,
-                            emailAddresses
-                        )
-                    )
+                    setBody(subscription)
                 }
                 return@runCatching SubscriptionResponse.ProcessNotificationResponse(response)
             }
@@ -110,7 +75,7 @@ class UploadDigestCountsSubscriptionMutationService(
     @Suppress("unused")
     fun unsubscribeUploadDigestCounts(
         subscriptionId: String
-    ): NotificationSubscriptionResult {
+    ): WorkflowSubscriptionResult {
         val url = workflowServiceConnection.buildUrl("/unsubscribe/uploadDigestCounts")
 
         return runBlocking {
