+++ pstatus-notifications-workflow-ktor/src/main/kotlin/workflow/toperrors/DataStreamTopErrorsNotificationWorkflow.kt
@@ -1,6 +1,6 @@
 package gov.cdc.ocio.processingnotifications.workflow.toperrors
 
-import gov.cdc.ocio.processingnotifications.model.WorkflowSubscription
+import gov.cdc.ocio.types.model.WorkflowSubscription
 import io.temporal.workflow.WorkflowInterface
 import io.temporal.workflow.WorkflowMethod
